type error fixes and change ban constants

+++ src/validation.cpp
@@ -3371,7 +3371,7 @@ static bool AcceptBlockHeader(const CBlockHeader& block, bool fProofOfStake, CVa
 }
 
 // Exposed wrapper for AcceptBlockHeader
-bool ProcessNewBlockHeaders(int32_t& nPoSTemperature, const std::vector<CBlockHeader>& headers, CValidationState& state, const CChainParams& chainparams, const CBlockIndex** ppindex)
+bool ProcessNewBlockHeaders(uint32_t& nPoSTemperature, const std::vector<CBlockHeader>& headers, CValidationState& state, const CChainParams& chainparams, const CBlockIndex** ppindex)
 {
     {
         LOCK(cs_main);
@@ -3384,7 +3384,7 @@ bool ProcessNewBlockHeaders(int32_t& nPoSTemperature, const std::vector<CBlockHe
                 *ppindex = pindex;
             }
             nPoSTemperature += header.nFlags & BLOCK_PROOF_OF_STAKE ? 1 : -POW_HEADER_COOLING;
-            nPoSTemperature = std::max(nPoSTemperature, 0);
+            nPoSTemperature = std::max((int)nPoSTemperature, 0);
         }
     }
     NotifyHeaderTip();
